# Use Jest for JS/React tests