--- a/src/sentry/static/sentry/app/__mocks__/translations.jsx
+++ b/src/sentry/static/sentry/app/__mocks__/translations.jsx
@@ -1,0 +1,13 @@
+export function getTranslations(language) {
+  return {
+    '': {
+      domain: 'the_domain',
+      lang: 'en',
+      plural_forms: 'nplurals=2; plural=(n != 1);'
+    }
+  };
+}
+
+export function translationsExist(language) {
+  return true; // no translations for tests
+}