--- a/src/sentry/static/sentry/app/utils/__mocks__/localStorage.jsx
+++ b/src/sentry/static/sentry/app/utils/__mocks__/localStorage.jsx
@@ -1,0 +1,16 @@
+let localStorageMock = function() {
+  let store = {};
+  return {
+    getItem: function(key) {
+      return store[key];
+    },
+    setItem: function(key, value) {
+      store[key] = value.toString();
+    },
+    clear: function() {
+      store = {};
+    }
+  };
+};
+
+export default localStorageMock();