--- a/src/sentry/static/sentry/app/utils/pendingChangeQueue.jsx
+++ b/src/sentry/static/sentry/app/utils/pendingChangeQueue.jsx
@@ -1,6 +1,5 @@
-class PendingChangeQueue extends Array {
+class PendingChangeQueue {
   constructor() {
-    super();
     this.changes = [];
   }