--- a/tests/js/setup.js
+++ b/tests/js/setup.js
@@ -1,0 +1,7 @@
+jest.mock('app/translations');
+
+import jQuery from 'jquery';
+window.$ = window.jQuery = jQuery;
+
+import sinon from 'sinon';
+window.sinon = sinon;