--- a/tests/js/spec/components/contextData.spec.jsx
+++ b/tests/js/spec/components/contextData.spec.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import TestUtils from 'react-addons-test-utils';
+import {mount} from 'enzyme';
 
 import ContextData from 'app/components/contextData';
 
@@ -9,12 +8,10 @@ describe('ContextData', function() {
     describe('strings', function() {
       it('should render urls w/ an additional <a> link', function() {
         const URL = 'https://example.org/foo/bar/';
-        let ctxData = TestUtils.renderIntoDocument(<ContextData data={URL} />);
+        let wrapper = mount(<ContextData data={URL} />);
 
-        let node = ReactDOM.findDOMNode(ctxData);
-
-        expect(node.getElementsByTagName('span')[0]).to.have.property('textContent', URL);
-        expect(node.getElementsByTagName('a')[0]).to.have.property('href', URL);
+        expect(wrapper.find('span').at(0).text()).toEqual(URL);
+        expect(wrapper.find('a').at(0).prop('href')).toEqual(URL);
       });
     });
   });