* `npm test` now uses Jest
* Assertions changed to use Jest's `expect` API
* Removes previous Mocha, Chai, Webpack, and PhantomJS launcher test runner / dependencies
* Faster (~21ms with no-cache vs ~27ms using Webpack/Phantom)
* Doesn't use snapshot testing (_yet_) – but can totally do so going forward
* Fixes some bugs caught by test changes

**Before:**

```bash
(sentry)➜  sentry git:(master) ✗ time npm test

> Sentry@0.0.0 test /Users/benvinegar/Projects/sentry
> node_modules/karma/bin/karma start tests/karma.conf.js --single-run

17 04 2017 21:41:23.777:INFO [karma]: Karma v0.13.22 server started at http://localhost:9876/
17 04 2017 21:41:23.792:INFO [launcher]: Starting browser PhantomJS
17 04 2017 21:41:25.628:INFO [PhantomJS 2.1.1 (Mac OS X 0.0.0)]: Connected on socket -_daL7ZwRenGFBUJAAAA with id 13279204
17 04 2017 21:41:26.328:WARN [web-server]: 404: /undefined/avatar/e1f3994f2632af3d1c8c2dcc168a10e6?s=48&d=blank
17 04 2017 21:41:26.383:WARN [web-server]: 404: /undefined/avatar/da34d4c0d7f41209cfee6eaf559905a4?s=48&d=blank
.................17 04 2017 21:41:26.435:WARN [web-server]: 404: /undefined/avatar/e1f3994f2632af3d1c8c2dcc168a10e6?s=48&d=blank
17 04 2017 21:41:26.526:WARN [web-server]: 404: /undefined/avatar/da34d4c0d7f41209cfee6eaf559905a4?s=48&d=blank
..........................................................
ERROR: 'Warning: Failed prop type: Invalid prop `identifier` of type `number` supplied to `LetterAvatar`, expected `string`.
    in LetterAvatar'
................................................................................
....................
ERROR: 'Warning: Unknown prop `pageLinks` on <div> tag. Remove this prop from the element. For details, see https://fb.me/react-unknown-prop
    in div (created by ProjectReleases)
    in div (created by ProjectReleases)
    in ProjectReleases'
.................
ERROR: 'Warning: Failed prop type: Invalid prop `message` of type `boolean` supplied to `LoadingError`, expected `string`.
    in LoadingError'
...............................
PhantomJS 2.1.1 (Mac OS X 0.0.0): Executed 223 of 223 SUCCESS (0.897 secs / 0.482 secs)
npm test  27.51s user 3.08s system 70% cpu 43.297 total
```

**After:**

```bash
(sentry)➜  sentry git:(use-jest) jest tests/js/spec --no-watchman
 PASS  tests/js/spec/components/assigneeSelector.spec.jsx (10.117s)
 PASS  tests/js/spec/views/stream.spec.jsx (11.144s)
  ● Console

    console.error node_modules/fbjs/lib/warning.js:36
      Warning: Failed prop type: Invalid prop `message` of type `boolean` supplied to `LoadingError`, expected `string`.
          in LoadingError

 PASS  tests/js/spec/views/groupActivity/index.spec.jsx
 PASS  tests/js/spec/components/events/interfaces/breadcrumbComponents/breadcrumbs.spec.jsx
 PASS  tests/js/spec/views/projectReleases.spec.jsx
  ● Console

    console.error node_modules/fbjs/lib/warning.js:36
      Warning: Unknown prop `pageLinks` on <div> tag. Remove this prop from the element. For details, see https://fb.me/react-unknown-prop
          in div (created by ProjectReleases)
          in div (created by ProjectReleases)
          in ProjectReleases

 PASS  tests/js/spec/views/stream/actions.spec.jsx
 PASS  tests/js/spec/components/events/interfaces/richHttpContent.spec.jsx
 PASS  tests/js/spec/views/stream/searchBar.spec.jsx
 PASS  tests/js/spec/components/issues/snoozeAction.spec.jsx
 PASS  tests/js/spec/views/releaseArtifacts.spec.jsx
 PASS  tests/js/spec/views/organizationTeams.spec.jsx
 PASS  tests/js/spec/views/projectInstall/platform.spec.jsx
 PASS  tests/js/spec/components/events/interfaces/keyValueList.spec.jsx
 PASS  tests/js/spec/components/events/interfaces/breadcrumbComponents/httpRenderer.spec.jsx
 PASS  tests/js/spec/api.spec.jsx
 PASS  tests/js/spec/components/eventsPerHour.spec.jsx
 PASS  tests/js/spec/views/stream/actionLink.spec.jsx
 PASS  tests/js/spec/components/avatarCropper.spec.jsx
 PASS  tests/js/spec/components/avatar.spec.jsx
 PASS  tests/js/spec/views/ruleEditor/ruleNodeList.spec.jsx
 PASS  tests/js/spec/components/dropdownLink.spec.jsx
 PASS  tests/js/spec/components/group/tagDistributionMeter.spec.jsx
 PASS  tests/js/spec/views/groupDetails/seenBy.spec.jsx
 PASS  tests/js/spec/components/letterAvatar.spec.jsx
  ● Console

    console.error node_modules/fbjs/lib/warning.js:36
      Warning: Failed prop type: Invalid prop `identifier` of type `number` supplied to `LetterAvatar`, expected `string`.
          in LetterAvatar

 PASS  tests/js/spec/components/barChart.spec.jsx (17.435s)
 PASS  tests/js/spec/components/contextData.spec.jsx
 PASS  tests/js/spec/components/events/interfaces/frame.spec.jsx
 PASS  tests/js/spec/components/events/interfaces/rawStacktraceContent.spec.jsx
 PASS  tests/js/spec/components/userLetterAvatar.spec.jsx
  ● Console

    console.error node_modules/fbjs/lib/warning.js:36
      Warning: Failed prop type: Invalid prop `identifier` of type `number` supplied to `LetterAvatar`, expected `string`.
          in LetterAvatar (created by UserLetterAvatar)
          in UserLetterAvatar

 PASS  tests/js/spec/stores/groupStore.spec.jsx
 PASS  tests/js/spec/components/events/interfaces/utils.spec.jsx
 PASS  tests/js/spec/utils/streamManager.spec.js
 PASS  tests/js/spec/utils/utils.spec.jsx
 PASS  tests/js/spec/utils/logging.spec.jsx
 PASS  tests/js/spec/stores/alertStore.spec.js
 PASS  tests/js/spec/stores/streamTagStore.spec.jsx
 PASS  tests/js/spec/utils/stream.spec.jsx
 PASS  tests/js/spec/utils/marked.spec.jsx
 PASS  tests/js/spec/helpers/formatters.spec.jsx
 PASS  tests/js/spec/stores/selectedGroupStore.spec.js

Test Suites: 40 passed, 40 total
Tests:       222 passed, 222 total
Snapshots:   0 total
Time:        21.416s
Ran all test suites matching "tests/js/spec".
```

**Notes:**

* I personally ran into Jest sometimes hanging / Watchman not responding. If you run into this, check out the [Watchman troubleshooting notes](https://facebook.github.io/jest/docs/troubleshooting.html#watchman-issues) from the Jest docs.