@benvinegar 2017-04-18
> Executed 223 of 223 SUCCESS (0.897 secs / 0.482 secs)
> Tests:       222 passed, 222 total

Oh yeah, I removed one test that literally just mounted a component and did nothing else.